edit : add page create on opertioncand edit url

+++ app/Http/Controllers/Operations/ClientController.php
@@ -72,6 +72,62 @@ class ClientController extends Controller
     {
         return $userDataTable->render('operations.client.index');
     }
+    public function create()
+    {
+
+        $rolesSelected = [];
+        $hasCustomField = in_array($this->userRepository->model(), setting('custom_field_models', []));
+        if ($hasCustomField) {
+            $customFields = $this->customFieldRepository->findByField('custom_field_model', $this->userRepository->model());
+            $html = generateCustomField($customFields);
+        }
+
+        return view('operations.client.create')
+            // ->with("role", $role)
+            ->with("customFields", isset($html) ? $html : false)
+            ->with("rolesSelected", $rolesSelected);
+    }
+     /**
+     * Store a newly created User in storage.
+     *
+     * @param CreateUserRequest $request
+     *
+     * @return Response
+     */
+    public function store(CreateUserRequest $request)
+    {
+        if (env('APP_DEMO', false)) {
+            Flash::warning('This is only demo app you can\'t change this section ');
+            return redirect(route('users.index'));
+        }
+
+        $input = $request->all();
+        $customFields = $this->customFieldRepository->findByField('custom_field_model', $this->userRepository->model());
+
+        $input['roles'] =  ['client'];
+        $input['password'] = Hash::make($input['password']);
+        $input['api_token'] = str_random(124);
+        $input['activated_at'] = now();
+
+        try {
+            $user = $this->userRepository->create($input);
+            $user->syncRoles($input['roles']);
+            $user->customFieldsValues()->createMany(getCustomFieldsValues($customFields, $request));
+
+            if (isset($input['avatar']) && $input['avatar']) {
+                $cacheUpload = $this->uploadRepository->getByUuid($input['avatar']);
+                $mediaItem = $cacheUpload->getMedia('avatar')->first();
+                $mediaItem->copy($user, 'avatar');
+            }
+            event(new UserRoleChangedEvent($user));
+        } catch (ValidatorException $e) {
+            Flash::error($e->getMessage());
+        }
+
+        Flash::success('saved successfully.');
+
+        return redirect(route('operations.users.index'));
+    }
       /**
      * Display the specified Order.
      *
@@ -230,7 +286,6 @@ class ClientController extends Controller
         $user = $this->userRepository->findWithoutFail($id);
         unset($user->password);
         $html = false;
-        $role = $this->roleRepository->pluck('name', 'name');
         $rolesSelected = $user->getRoleNames()->toArray();
         $customFieldsValues = $user->customFieldsValues()->with('customField')->get();
         $customFields = $this->customFieldRepository->findByField('custom_field_model', $this->userRepository->model());
@@ -259,7 +314,8 @@ class ClientController extends Controller
 
 
         return view('operations.client.edit')
-            ->with('user', $user)->with("role", $role)
+            ->with('user', $user)
+            // ->with("role", $role)
             ->with("rolesSelected", $rolesSelected)
             ->with("customFields", $html);
     }
@@ -276,11 +332,11 @@ class ClientController extends Controller
     {
         if (env('APP_DEMO', false)) {
             Flash::warning('This is only demo app you can\'t change this section ');
-            return redirect(route('users.profile'));
+            return redirect()->back();
         }
         if (!auth()->user()->hasRole('admin') && $id != auth()->id()) {
             Flash::error('Permission denied');
-            return redirect(route('users.profile'));
+            return redirect()->back();
         }
 
         $user = $this->userRepository->findWithoutFail($id);
@@ -330,9 +386,10 @@ class ClientController extends Controller
         }
 
 
-        Flash::success('User updated successfully.');
+        Flash::success($user->name.'updated successfully.');
+        return redirect(route('operations.users.index'));
 
-        return redirect()->back();
+        // return redirect()->back();
 
     }
 
@@ -354,8 +411,11 @@ class ClientController extends Controller
 
         if (empty($user)) {
             Flash::error('User not found');
-
-            return redirect(route('users.index'));
+            return redirect(route('operations.users.index'));
+        }
+        if ($user->orders()->exists()) {
+            Flash::error($user->name.trans('lang.error_user_have_orders'));
+            return redirect(route('operations.users.index'));
         }
 
         $this->userRepository->delete($id);
